[FL-572][FL-577] Irda receive feature (#282)
* fix "state not acquired error"
* add InterruptTypeComparatorTrigger to interrupt mgr, use interrupt mgr in irda app
* separate init irda timer
* capture events buffer by app
* irda common decoder
* irda nec decoder realization
* finished work with decoder
* fix app path
* fix widget remove on exit
* nec receive, store and send
* init some packets

+++ applications/irda/irda_nec.h
@@ -1,4 +1,4 @@
 #pragma once
 #include "flipper.h"
 
-void ir_nec_send(uint8_t addr, uint8_t data);
+void ir_nec_send(uint16_t addr, uint8_t data);
